+++ src/main/java/com/restful/RestFullDemoApplication.java
@@ -1,4 +1,4 @@
-package com.restfulldemo;
+package com.restful;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
