+++ src/main/java/com/restful/exception/NotParsableContentException.java
@@ -1,4 +1,4 @@
-package com.restfulldemo.exception;
+package com.restful.exception;
 
 public class NotParsableContentException extends Exception{
