+++ src/main/java/com/restful/model/Link.java
@@ -1,4 +1,4 @@
-package com.restfulldemo.model;
+package com.restful.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
